Refatoração e adição de controladores no API
Renomeado `CreateResponsavelRequest` para `CreateClientRequest` para melhor clareza.
Alterada a propriedade `CostCenterId` de `int?` para `long?` em `PaymentPlan.cs` para suportar valores maiores.
Adicionado o pacote `Microsoft.EntityFrameworkCore.SqlServer` no projeto para suporte ao SQL Server.
Implementados os controladores `ChargesController`, `ClientsController` e `PlansController` para gerenciar cobranças, clientes e planos de pagamento, respectivamente.
Reorganizada a configuração do banco de dados em `Program.cs` para permitir a escolha entre SQL Server e PostgreSQL.

+++ PaymentPlanManagement_API/Domain/Entities/PaymentPlan.cs
@@ -8,7 +8,7 @@ public class PaymentPlan
 
     public Client Client { get; private set; } = null!;
 
-    public int? CostCenterId { get; private set; }
+    public long? CostCenterId { get; private set; }
 
     public CostCenter? CostCenter { get; private set; }
 
@@ -24,7 +24,7 @@ public class PaymentPlan
 
     protected PaymentPlan() { }
 
-    public PaymentPlan(long clientId, int? costCenterId, string? costCenterEnum)
+    public PaymentPlan(long clientId, long? costCenterId, string? costCenterEnum)
     {
       ClientId = clientId;
       CostCenterId = costCenterId;
